Initial 64-bit Windows.

+++ native/ax/Native2.0/NativeCore/NativeCore.c
@@ -232,8 +232,8 @@ char* _access_alignment_check_( char* buf, int buf_len )
 	char* b_c_p  = (char*)&b.c;
 	//char* b_d_p  = (char*)&b.d;
 	snprintf( buf, buf_len, "a: %ld %ld %ld %ld; b:%ld %ld", 
-		      a_b_p-base_a, a_c_p-base_a, a_d_p-base_a, a_e_p-base_a, 
-			    b_b_p-base_b, b_c_p-base_b );
+		      (int) (a_b_p-base_a), (int) (a_c_p-base_a), (int) (a_d_p-base_a), (int) (a_e_p-base_a), 
+			    (int) (b_b_p-base_b), (int) (b_c_p-base_b) );
 
 	return buf;
